requirements.txt
customtkinter
requests
websocket-client
Pillow
pyinstaller
